[Gold III] Title: 두 배열의 합, Time: 232 ms, Memory: 41076 KB -BaekjoonHub

+++ 백준/Gold/2143. 두 배열의 합/두 배열의 합.cc
@@ -35,7 +35,7 @@ int main(void) {
 		}
 	}
 
-	long long answer = 0;
+	ll answer = 0;
 	for (auto& p : possible_sum_count_A) {
 		int target_sum = T - p.first;
 		if (possible_sum_count_B.find(target_sum) == possible_sum_count_B.end()) continue;
